# openapi.yaml
openapi: 3.0.0
paths:
  /:
    get:
      description: Returns API information and available endpoints
      operationId: RootController_getRoot
      parameters: []
      responses:
        '200':
          description: API information retrieved successfully
      summary: API Information
      tags:
        - root
  /users:
    get:
      description: Retrieves a list of all users in the system
      operationId: AppController_getUsers
      parameters: []
      responses:
        '200':
          description: List of users successfully retrieved
      summary: Get all users
      tags: &ref_0
        - users
  /users/{id}:
    get:
      description: Retrieves a specific user by their unique identifier
      operationId: AppController_getUser
      parameters:
        - name: id
          required: true
          in: path
          description: User id
          schema:
            type: number
      responses:
        '200':
          description: User found
      summary: Get a user by id
      tags: *ref_0
  /products:
    get:
      description: Retrieves a list of all available products in the catalog
      operationId: getAllProducts
      parameters: []
      responses:
        '200':
          description: List of products successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    price:
                      type: number
        '500':
          description: Internal server error
      summary: Get all products
      tags: &ref_1
        - products
  /products/{id}:
    get:
      operationId: getProductById
      parameters:
        - name: id
          required: true
          in: path
          description: Product unique identifier
          schema:
            type: number
      responses:
        '200':
          description: Product found
      summary: Get product by id
      tags: *ref_1
    put:
      description: Updates an existing product by its identifier
      operationId: updateProduct
      parameters:
        - name: id
          required: true
          in: path
          description: Product unique identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
      responses:
        '200':
          description: Product updated successfully
        '400':
          description: Invalid input data
        '404':
          description: Product not found
      summary: Update a product
      tags: *ref_1
    delete:
      description: Deletes a product from the catalog
      operationId: ProductsController_deleteProduct
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: Product deleted successfully
      summary: Delete product
      tags: *ref_1
  /products/createProduct:
    post:
      description: Creates a new product in the catalog
      operationId: createProduct
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
              required:
                - name
                - price
      responses:
        '201':
          description: Product created successfully
        '400':
          description: Invalid input data
      summary: Create a new product
      tags: *ref_1
  /orders:
    get:
      description: Retrieves a list of all orders with optional filtering
      operationId: getAllOrders
      parameters:
        - name: status
          required: false
          in: query
          description: Filter orders by status
          schema:
            enum:
              - pending
              - completed
              - cancelled
            type: string
      responses:
        '200':
          description: List of orders successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    userId:
                      type: number
                    status:
                      type: string
                    total:
                      type: number
        '400':
          description: Invalid filter parameters
        '500':
          description: Internal server error
      summary: Get all orders
      tags: &ref_2
        - orders
    post:
      description: Creates a new order for a user
      operationId: createOrder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: number
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: number
                      quantity:
                        type: number
              required:
                - userId
                - items
      responses:
        '201':
          description: Order created successfully
      summary: Create a new order
      tags: *ref_2
  /orders/{id}:
    get:
      description: Retrieves a specific order by its unique identifier
      operationId: getProductById
      parameters:
        - name: id
          required: true
          in: path
          description: Order unique identifier
          schema:
            type: number
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  userId:
                    type: number
                  status:
                    type: string
                  total:
                    type: number
        '404':
          description: Order not found
      summary: Get order by id
      tags: *ref_2
  /orders/user_orders/{userId}:
    get:
      description: Retrieves all orders for a specific user
      operationId: getUserOrders
      parameters:
        - name: userId
          required: true
          in: path
          description: User unique identifier
          schema:
            type: number
      responses:
        '200':
          description: User orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      summary: Get orders by user
      tags: *ref_2
  /orders/{id}/cancel:
    post:
      description: Cancels an existing order
      operationId: cancelOrder
      parameters:
        - name: id
          required: true
          in: path
          description: Order unique identifier
          schema:
            type: number
      responses:
        '200':
          description: Order cancelled successfully
        '400':
          description: Order cannot be cancelled
        '404':
          description: Order not found
      summary: ''
      tags: *ref_2
info:
  title: Backend Microservice Infra Boilerplate
  description: Backend Microservice Infra Boilerplate
  version: 1.0.0
  contact: {}
tags:
  - name: users
    description: User management endpoints
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  schemas: {}